move EngineCoreReadyResponse definition back
