+++ vllm/v1/engine/__init__.py
@@ -4,6 +4,7 @@
 import enum
 import time
 from collections.abc import Mapping
+from dataclasses import dataclass
 from typing import Any, Literal
 
 import msgspec
@@ -61,6 +62,19 @@ class FinishReason(enum.IntEnum):
 
     def __str__(self):
         return FINISH_REASON_STRINGS[self.value]
+
+
+@dataclass
+class EngineCoreReadyResponse:
+    """Sent from EngineCore to each frontend at the end of engine startup.
+
+    Contains post-initialization config that may differ from the original
+    values (e.g. max_model_len after KV cache auto-fitting).
+    """
+
+    num_gpu_blocks: int
+    dp_stats_address: str | None
+    max_model_len: int | None = None
 
 
 class EngineCoreRequest(
